nand fix
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@3845 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/linux/rb500/linux-2.6/drivers/ide/pci/hpt34x.c
@@ -24,6 +24,7 @@
  * Non-bootable card or HPT343 :: pcicmd == 0x05
  */
 
+#include <linux/config.h>
 #include <linux/module.h>
 #include <linux/types.h>
 #include <linux/kernel.h>
@@ -175,7 +176,7 @@ static unsigned int __devinit init_chipset_hpt34x(struct pci_dev *dev, const cha
 			pci_write_config_dword(dev, PCI_ROM_ADDRESS,
 				dev->resource[PCI_ROM_RESOURCE].start | PCI_ROM_ADDRESS_ENABLE);
 			printk(KERN_INFO "HPT345: ROM enabled at 0x%08lx\n",
-				(unsigned long)dev->resource[PCI_ROM_RESOURCE].start);
+				dev->resource[PCI_ROM_RESOURCE].start);
 		}
 		pci_write_config_byte(dev, PCI_LATENCY_TIMER, 0xF0);
 	} else {
